Fixed AutocompleteSearch tests, created ApiClient
